añadidos test para nuevos endpoints

+++ API-Inmobiliaria/src/main/java/org/inmobiliaria/apiinmobiliaria/repository/InmuebleRepository.java
@@ -14,5 +14,5 @@ public interface InmuebleRepository extends CrudRepository<Inmueble, Long> {
     List<Inmueble> findAll();
 
     @Query("SELECT i FROM Inmueble i WHERE i.precio BETWEEN :min AND :max")
-    List<Inmueble> findInmueblesRangoPrecio(Double min, Double max);
+    List<Inmueble> findInmueblesRangoPrecio(Float min, Float max);
 }
